// src/app/shared/models/people.interface.tsx
export interface People {
  birth_year: string
  created: string
  edited: string
  eye_color: string
  films: Array<string>
  gender: string
  hair_color: string
  height: string
  homeworld: string
  mass: string
  name: string
  skin_color: string
  species: Array<string>
  starships: Array<string>
  url: Array<string>
  vehicles: Array<string>
}